requirements.txt
amqp==5.3.1
asgiref==3.8.1
bcrypt==4.3.0
billiard==4.2.1
boto3==1.35.95
botocore==1.35.99
celery==5.5.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cryptography==44.0.2
Django==5.1.7
django-celery-beat==2.7.0
django-ckeditor==6.7.2
django-js-asset==3.1.2
django-storages==1.14.6
django-timezone-field==7.1
dnspython==2.7.0
eventlet==0.39.1
ghasedak_sms==1.0.3
greenlet==3.1.1
idna==3.10
jmespath==1.0.1
kavenegar==1.1.2
kombu==5.5.2
paramiko==3.5.1
pillow==11.1.0
prompt_toolkit==3.0.50
psycopg==3.2.6
psycopg-binary==3.2.6
psycopg-pool==3.2.6
pycparser==2.22
PyNaCl==1.5.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
requests==2.32.3
s3transfer==0.10.4
six==1.17.0
sqlparse==0.5.3
storage==0.0.4.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
